move gif alt text dialog for consistency
