+++ src/view/com/composer/photos/GifAltTextDialog.tsx
